Adding creep wrappers, base, worker and miner
This is a restore of code backup from 13.12.2016.

Have decided to use wrappers instead of manipulating game objects.
Started with a wrapper for the Miner. Using a base class for common logic. Worker class not in use.
Other roles still defined in main using task logic for harvesting and delivering in separate modules.
Code versioning for initialization of memory.